+++ CHANGELOG.md
@@ -8,10 +8,14 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- Enable parent/children filters in the filters pane in Search.
+
 ### Changed
 
 - Translations are loaded dynamically in frontend application.
-- Optimized frontend build in our official Docker image
+- Optimized frontend build in our official Docker image.
 
 ## [1.0.0-beta.4] - 2019-04-08